requirements.txt
numpy
pandas
statsmodels
scikit-learn
scikit-optimize
matplotlib
jupyter
wandb
plotly
torch
torchvision
pillow
albumentations
seaborn